+++ server/gen-graphql-types.ts
@@ -4,7 +4,7 @@ import { resolve } from 'path';
 generate(
   {
     require: ['ts-node/register/transpile-only'],
-    schema: resolve(__dirname, 'src/schema/index.ts'),
+    schema: resolve(__dirname, 'src/schema/schema.ts'),
     template: 'graphql-codegen-typescript-template',
     out: resolve(__dirname, 'src/graphql-types.ts'),
     overwrite: true
