+++ server/src/schema/index.ts
@@ -1,6 +1,6 @@
 import { makeExecutableSchema } from 'apollo-server-express';
 import { resolvers } from './resolvers';
-import { typeDefs } from './schema';
+import { typeDefs } from './type-defs';
 
 export const schema = makeExecutableSchema({
   resolvers: resolvers as any,
